Refactor: Inline EncodedImageMap into DesignComposeDefinition (#1845)
<!-- start git-machete generated -->

## Tree of downstream PRs as of 2024-12-09

* **PR #1845 (THIS ONE)**:
  `main` ← `wb/froeht/convert-encodedimagemap`

    * PR #1847:
      `wb/froeht/convert-encodedimagemap` ← `wb/froeht/1172-convert-dcd`

      * PR #1849:
        `wb/froeht/1172-convert-dcd` ← `wb/froeht/cargo-fix`

<!-- end git-machete generated -->

This change refactors the way images are stored in the
DesignComposeDefinition proto.
Instead of using separate EncodedImageMap and EncodedImage messages, the
`images` field is now a direct map of image keys to byte arrays.
This simplification removes the need for extra message types and makes
the definition more straightforward.

The existing EncodedImageMap has been promoted to the main definition
file. It will be used in the code and then cloned to the Proto message
before serialization.