Support autocontent preview widget in squoosh (#1852)
Fixes #1786. When the widget specifies a horizontal or vertical layout,
the code takes a path similar to that of ReplacementContent. The
children of the widget are replaced with the children specified in the
customization by placing them into the composable list. This handles
special content such as the overflow node.
    
When the widget specifies a grid layout, the code adds a single
composable to the composable list, a LazyVerticalGrid or
LazyHorizontalGrid. These handle their own layout and we reuse the
non-squoosh code.
    
This change removes the GridLayoutDocumentation example from the list of
examples in Validation since other tests already cover everything that
it does. The example still exists in case we want to modify it for use
in our documentation on how to use the list widget.
    
This change also moves the recursion in resolveVariantsRecursively() on
children to be after replacement component/content/list has been
resolved so that we don't recurse on children that will be removed
because of the replacement.