Temporarily disable the VariantAnimationTimelineTest. (#1950)
It tries to access parts of the DesignDoc that aren't meant to be
accessed, and fails to compile because Validation does not have the
protobuf libraries as a dependency. We don't want to expose the Protobuf
API, because this can lead to compilation problems when another library
uses a different version of the Protobuf library. We'll need to change
the test to resolve the compilation error without modifying it's Gradle
dependencies.

Task to fix it is #1945