Fix: Improve test setup for DocSerializationTest (#1964)
<!-- start git-machete generated -->

# Based on PR #1886

## Chain of upstream PRs as of 2025-01-09

* PR #1886:
  `main` ← `feature/protoconv`

  * **PR #1964 (THIS ONE)**:
    `feature/protoconv` ← `wb/froeht/improve-docserializationtest`

<!-- end git-machete generated -->

This change modifies the test setup for `DocSerializationTest` to
improve its robustness and coverage.

- Removes `showStandardStreams = true` from the roborazzi gradle config,
to quiet down the amount of console spam
- Adds the validation assets directory to the test resources for the
common module.
- Replaces the single `loadSaveLoadHelloWorld` test with a parameterized
test `LoadSaveLoadAllDocsTest` to make it easier to see which docs
aren't passing
- Removes the
`VariantAnimationTimelineTestDoc_vJRf4zxY4QX4zzSSUd1nJ5.dcf` file. The
test is disabled right now ( #1945 ) so the dcf wasn't being updated and
was failing to decode.