Fix TapCallback on nodes that change from an interaction (#1996)
Fixes #1995. When a node has both an interaction that changes the
variant and a TapCallback, the TapCallback was not being found because
it was customized using the node's original name. This has been fixed
with the following changes:

1. Add an isPressed mutableState that gets set to true when the press
happens, and false when released. This causes a recomposition which
builds a new UI tree using the new variant that has changed.
2. Add a set of pressed node IDs in interaction set so that when the
above recomposition happens, the new UI tree with the changed variant
still finds a supported interaction and adds itself to the root node's
composable child list.