Make Design Switcher functional again
Fixes #1881. The design switcher has been broken in squoosh due to a number of bugs:
https://github.com/google/automotive-design-compose/issues/1717
https://github.com/google/automotive-design-compose/issues/2000
https://github.com/google/automotive-design-compose/issues/2001
https://github.com/google/automotive-design-compose/issues/2002

Fixing all of these will take some time. In an effort to make the design switcher functional again, this change makes some small changes in the Figma design by removing some animations and interactions that are buggy. A few small changes in the design switcher code fix the functionality to switch between other documents or branches.