Support replacement component of a text node
Fixes #2001. Check for component replacement customizations on text nodes. When one exists, skip the text rendering path, and don't return early so that the replacement component is rendered. Save the TextStyle of the text being replaced and use it in the ComponentReplacementContext.

This change adds a new test for text replacement with a BasicTextField that inserts some text with keyboard events.