Switch Dependabot to run on main, better consolidate updates
The current Dependabot scheme to run on a "dependapool"
branch which would be merged required some manual steps.
GitHub's configured to delete branches after a merge
which meant I had to manually recreate the "dependapool"
branch each week.

I'm switching this back to run on main to avoid the manual work.

I'm also changing the configuration to better group the updates.
I realized that setting it to "minor" version updates still
created individual updates for "patch" releases. I'm adding a second
group to each for "major" updates (which are more likely to be breaking)

With this there'll still be up to ten dependabot CL's a week, (being generated on Fridays)
but usually less, because not all package ecosystems will have updates.