Fix overlays to be positioned relative to screen size again (#2018)
Fixes #1717: Instead of rendering overlays as a child of the root node,
render them after the root node as a composable inside of a full screen
box. This allows the overlay to be positioned relative to the size of
the entire screen instead of the root node. Use the overlay extras to
color the box and make the box clickable if that option is turned on.