Fix animation bugs (#2065)
Fixes #2000. Fix an issue where releasing a touch event while in the
middle of an onPress animation results in the variant getting stuck in
the onPress variant state instead of going back. Fix another issue where
finding the target node of a variant change animation would use an
existing node in the tree if it existed, resulting in an incorrect
layout animation to that node.