Fix the publish to localDir tasks to publish to a single dir
The `publishAllPublicationsToLocalDirRepository` task is intended to
create a single maven repository holding the entire DesignCompose
release. This was the case if the `DesignComposeMavenRepo` property
was set, but if it was unset it would create individual maven repos
in each of the subprojects.

This change will change that default to publish to the root project's
build directory instead. Running the task without the property set will
result in the repo being created in `build/designcompose_m2repo`