Add Gradle daemon JVM toolchain URLs for various OS and architectures (#2115)
This is a new Gradle 8.13 feature. To support it we needed to
re-generate the daemon-jvm file using `./gradlew updateDaemonJvm`

This commit introduces the `toolchainUrl` property to the
`gradle-daemon-jvm.properties` file. These properties provide specific
URLs for downloading the required Java toolchain for the Gradle daemon
on different operating systems and architectures, including:

-   FreeBSD (AARCH64, X86_64)
-   Linux (AARCH64, X86_64)
-   macOS (AARCH64, X86_64)
-   UNIX (AARCH64, X86_64)
-   Windows (AARCH64, X86_64)

The toolchain version is set to 21. These URLs will be used for
auto-provisioning the JVM required by the Gradle Daemon to run.