Support scrolling when dragging nodes with press/tap interactions (#2086)
Use the awaitEachGesture function of PointerInputScope to manually
handle touch events. This allows the touch gestures to propagate to the
parent, but a valid click will consume the event so as not to also click
the parent.

This change also fixes some scrolling bugs.
- Fixes an issue where the amount scrolled did not take into account
density. T
- Fixes an issue where the children of a scrollable frame were being
composed as children of the root node instead of the scrollable frame.
- Fixes an issue where nodes with press/tap interactions were rendered
scrolled, but their tappable areas were not, so tapping would dispatch
interactions on the wrong node.