Upgrade all self-hosted ubuntu-22.04 runners to ubuntu-24.04. (#2139)
This is a http://go/LSC run by http://go/ghss to upgrade all self-hosted
ubuntu-22.04 runners to ubuntu-24.04.

This is a courtesy PR to help you upgrade to the latest release of
ubuntu runners, it is not mandatory at this time.

WARNING: We do not know if the updated label is compatiable with your
workflow or not.

If you do not want to merge this PR, feel free to close it.

More context and feedback: http://b/406537467

---------

Co-authored-by: Bill Napier <napier@google.com>
Co-authored-by: yiqunw700 <yiqunw@google.com>