Fix issue that prevented use of DC in an includedBuild (#2164)
The publishing config (which tries to make it so the plugin get
published to the same directory as the main project) was checking to see
if the rootProject had any parents and made assumptions off of that.
Those assumptions weren't valid if the root DesignCompose Project as an
includedBuild of another project.

I've changed the checks to instead check for the rootProject's name, and
added IfChange guards to make sure that someone renaming one of the
projects doesn't break anything.