Increase compatibility with soong
Downgrade protobuf to 3.2.0:
Using the same version currently supported on Android to increase
compatilibty.

Use the pure Rust proto generator:
This removes dependency on the protoc binary to be on the PATH.

Use the default folder for proto genearation:
This works better with the Android crate import tooling.

Update thiserror to 2.0.11