Refactor to help importing the crate to AOSP (#2191)
Downgrade protobuf version to 3.2.0, the same version currently
supported on Android to reduce the number of patches required.
Move to pure rust code generator to avoid requiring protoc installed on
the host machine.
Drop using the "protos" subfolder, it is hard to work with in soong. The
proto files are included in the root 'lib.rs' anyway.
Upgrade thiserror to 2.x.

Fixes https://github.com/google/automotive-design-compose/issues/2195