Test for backward compatibility of old dcf files
Fixes #2126. This commit makes a change to the field containing a text node's font color, renaming it from text_color to font_color. Any updates to or new DCF files will save text color to font_color in node_style instead of text_color. This change's only purpose is to introduce a protobuf change and demonstrate our capability to support old protobufs that still set text_color. This is supported by using font_color if it is set, and if not, using text_color.

To facilitate support of old DCF files as protobufs continue to evolve, a new utility file utils/ProtoVersions.kt has some functions to manage checking of old fields. This file should grow when a protobuf change that requires code to support old versions is made.

Additionally, a new test has been added to test for old protobufs. Currently this tests that v27 protobufs (that still use text_color) have a header that says v27, and that they render correctly, including a screenshot. New tests should be added when protobufs again change.