feat: Upgrade to Kotlin 2.0 and Jetpack Compose Compiler 2.0 (#2232)
This change upgrades the project to Kotlin 2.0 and the new Jetpack
Compose compiler plugin.


  Note that a larger migration is required to support KSP 2: #2231

  ---

  Key Changes:


   * Upgraded Kotlin to version 2.1.21
* Replaced the AndroidX Compose compiler with the JetBrains Compose
compiler plugin version 2.1.21
* Added ksp.useKSP2=false to gradle.properties to maintain compatibility
with KSP 1.0 for now.
* Updated various build.gradle.kts files to use the new JetBrains
Compose plugin.
* Refactored the DesignCompose Gradle plugin to make it compatible with
the classloader isolation introduced in Kotlin 2.0. This was necessary
because the Kotlin 2.0 Gradle Plugin no longer includes the full
compiler on the build script classpath, which can cause issues for
plugins that depend on the Android Gradle Plugin (AGP). To address this,
all AGP-dependent code was moved into a new AndroidPluginHelper.kt file,
which is only loaded when the AGP plugin is present.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: farivar-google <farivar@google.com>