Fix #2243: ignore SECTION node type after parsing.
The parser generates a Section {} node data type for SECTION nodes now, which is then ignored during view tree construction. Generally components and DesignCompose nodes are inside of sections, so the section being ignored by visit_node doesn't make any difference.

If a SECTION somehow does appear inside of a node that forms a composable, then it'll be treated as a default rectangle with children.